Time: 346 ms (5%), Space: 45.3 MB (12.94%) - LeetHub

+++ 0018-4sum/0018-4sum.java
@@ -5,11 +5,11 @@ class Solution {
         Arrays.sort(nums);
         for(int i=0;i<n-3;i++){
             for(int j=i+1;j<n-2;j++){
-                int sum = nums[i]+nums[j];
+                long sum = (long)nums[i]+ (long)nums[j];
                 int s=j+1;
                 int e=n-1;
                 while(s<e){
-                    long tempSum = sum+nums[s]+nums[e];
+                    long tempSum = sum+ (long)nums[s]+ (long)nums[e];
                     if(tempSum == target){
                         if(!ans.contains(Arrays.asList(nums[i],nums[j],nums[s],nums[e]))){
                             ans.add(Arrays.asList(nums[i],nums[j],nums[s],nums[e]));
